+++ src/lib/zodSchemas.ts
@@ -1,3 +1,4 @@
+import { evaluate } from 'mathjs';
 import { z } from 'zod';
 
 export const exerciseEventSchema = z.object({
@@ -26,3 +27,77 @@ export const calendarEventSchema = z.object({
 export const calendarEventPartialSchema = calendarEventSchema.partial();
 export type CalendarEventSchema = typeof calendarEventSchema;
 export type CalendarEventPartialSchema = typeof calendarEventPartialSchema;
+
+export const customQuerySchema = z
+  .object({
+    name: z.string().min(1, { message: 'Name is required' }),
+    table: z.enum(['metric', 'exerciseEvent']).default('exerciseEvent'),
+    equation: z.string().min(1, { message: 'Equation is required' }),
+    metric: z.string().nullish(),
+    exerciseId: z.string().nullish(),
+  })
+  .superRefine((val, ctx) => {
+    if (val.table == 'exerciseEvent' && !val.exerciseId) {
+      ctx.addIssue({
+        code: z.ZodIssueCode.custom,
+        message: `Exercise is required`,
+        path: ['exercise'],
+      });
+    }
+    if (val.table == 'metric' && !val.metric) {
+      ctx.addIssue({
+        code: z.ZodIssueCode.custom,
+        message: `Metric is required`,
+        path: ['metric'],
+      });
+    }
+    // Check if equation is valid by attempting to solve the equation
+    try {
+      if (val.table == 'exerciseEvent') {
+        evaluate(val.equation, {
+          sets: 0,
+          reps: 0,
+          weight: 0,
+          minutes: 0,
+          seconds: 0,
+        });
+      } else if (val.table == 'metric') {
+        evaluate(val.equation, {
+          value: 0,
+        });
+      }
+    } catch (e: any) {
+      ctx.addIssue({
+        code: z.ZodIssueCode.custom,
+        message: e.toString(),
+        path: ['equation'],
+      });
+    }
+  });
+export type CustomQuerySchema = typeof customQuerySchema;
+
+export const customQueryConditionSchema = z
+  .object({
+    column: z.string().min(1, { message: 'Column is required' }),
+    condition: z.enum(['equals', 'contains']).default('contains'),
+    useWidgetField: z.boolean().default(false).optional(),
+    widgetFieldToUse: z.string().nullish(),
+    value: z.number(),
+  })
+  .superRefine((val, ctx) => {
+    // If useWidgetField is set to false, set the widgetFieldToUse to null
+    // so that we don't have a potential dangling reference to a field on the
+    // widget that may possibly have been removed by another action
+    if (val.useWidgetField === false) {
+      val.widgetFieldToUse = null;
+    }
+
+    if (val.useWidgetField === true && val.widgetFieldToUse === null) {
+      ctx.addIssue({
+        code: z.ZodIssueCode.custom,
+        message: 'Must specify a widget field to use',
+        path: ['widgetFieldToUse'],
+      });
+    }
+  });
+export type CustomQueryConditionSchema = typeof customQueryConditionSchema;
